Move ABResult to ABChecker
Moved the ABResult to be inside ABChecker.
